chore: new package for file collection
