+++ pkg/test/case.go
@@ -3,11 +3,7 @@ package test
 import (
 	"context"
 	"fmt"
-	"os"
-	"path/filepath"
-	"regexp"
 	"sort"
-	"strconv"
 	"testing"
 	"time"
 
@@ -26,10 +22,8 @@ import (
 	"github.com/kudobuilder/kuttl/pkg/report"
 	testutils "github.com/kudobuilder/kuttl/pkg/test/utils"
 	eventutils "github.com/kudobuilder/kuttl/pkg/test/utils/events"
+	"github.com/kudobuilder/kuttl/pkg/test/utils/files"
 )
-
-// testStepRegex contains one capturing group to determine the index of a step file.
-var testStepRegex = regexp.MustCompile(`^(\d+)-(?:[^\.]+)(?:\.yaml)?$`)
 
 // Case contains all of the test steps and the Kubernetes client and other global configuration
 // for a test.
@@ -269,66 +263,9 @@ func (t *Case) determineNamespace() (*namespace, error) {
 	return ns, nil
 }
 
-// CollectTestStepFiles collects a map of test steps and their associated files
-// from a directory.
-func (t *Case) CollectTestStepFiles() (map[int64][]string, error) {
-	testStepFiles := map[int64][]string{}
-
-	files, err := os.ReadDir(t.Dir)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, file := range files {
-		index, err := getIndexFromFile(file.Name())
-		if err != nil {
-			return nil, err
-		}
-		if index < 0 {
-			t.Logger.Log("Ignoring", file.Name(), "as it does not match file name regexp:", testStepRegex.String())
-			continue
-		}
-
-		if testStepFiles[index] == nil {
-			testStepFiles[index] = []string{}
-		}
-
-		testStepPath := filepath.Join(t.Dir, file.Name())
-
-		if file.IsDir() {
-			testStepDir, err := os.ReadDir(testStepPath)
-			if err != nil {
-				return nil, err
-			}
-
-			for _, testStepFile := range testStepDir {
-				testStepFiles[index] = append(testStepFiles[index], filepath.Join(
-					testStepPath, testStepFile.Name(),
-				))
-			}
-		} else {
-			testStepFiles[index] = append(testStepFiles[index], testStepPath)
-		}
-	}
-
-	return testStepFiles, nil
-}
-
-// getIndexFromFile returns the index derived from fileName's prefix, ex. "01-foo.yaml" has index 1.
-// If an index isn't found, -1 is returned.
-func getIndexFromFile(fileName string) (int64, error) {
-	matches := testStepRegex.FindStringSubmatch(fileName)
-	if len(matches) != 2 {
-		return -1, nil
-	}
-
-	i, err := strconv.ParseInt(matches[1], 10, 32)
-	return i, err
-}
-
 // LoadTestSteps loads all of the test steps for a test case.
 func (t *Case) LoadTestSteps() error {
-	testStepFiles, err := t.CollectTestStepFiles()
+	testStepFiles, err := files.CollectTestStepFiles(t.Dir, t.Logger)
 	if err != nil {
 		return err
 	}
