+++ pkg/test/utils/files/test_data/with-overrides/00-assert.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test
+status:
+  qosClass: BestEffort
