+++ pkg/test/utils/files/test_data/with-overrides/00-test-step.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test
+spec:
+  restartPolicy: Never
+  containers:
+  - name: nginx
+    image: nginx:1.7.9
+---
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+metadata:
+  name: with-test-step-name-override
