+++ pkg/test/utils/files/test_data/with-overrides/01-assert.yaml
@@ -1,0 +1,10 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestAssert
+timeout: 20
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test2
+status:
+  qosClass: BestEffort
