+++ pkg/test/utils/files/test_data/with-overrides/01-test-assert.yaml
@@ -1,0 +1,16 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test2
+spec:
+  restartPolicy: Never
+  containers:
+  - name: nginx
+    image: nginx:1.7.9
+---
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+delete:
+- apiVersion: v1
+  kind: Pod
+  name: test
