+++ pkg/test/utils/files/test_data/with-overrides/02-directory/assert.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test3
+status:
+  qosClass: BestEffort
