+++ pkg/test/utils/files/test_data/with-overrides/02-directory/pod.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test4
+spec:
+  containers:
+  - name: nginx
+    image: nginx:1.7.9
