+++ pkg/test/utils/files/test_data/with-overrides/03-assert.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test5
+spec:
+  restartPolicy: Never
