+++ pkg/test/utils/files/test_data/with-overrides/03-pod.yaml
@@ -1,0 +1,14 @@
+apiVersion: kuttl.dev/v1beta1
+kind: TestStep
+metadata:
+  name: name-overridden
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test6
+spec:
+  restartPolicy: Never
+  containers:
+  - name: nginx
+    image: nginx:1.7.9
