+++ pkg/test/utils/files/test_data/with-overrides/03-pod2.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: test5
+spec:
+  restartPolicy: Never
+  containers:
+  - name: nginx
+    image: nginx:1.7.9
